cxr-audit-api/api_requirements.txt
# ================================
# FastAPI Server Dependencies
# ================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.0.0

# ================================
# Data Processing & Analysis
# ================================
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
statsmodels>=0.13.0

# ================================
# File I/O & Excel Processing
# ================================
openpyxl>=3.0.0
xlrd>=2.0.0
msoffcrypto-tool>=5.0.0

# ================================
# Machine Learning & LLM Integration
# ================================
openai>=1.0.0

# ================================
# Data Visualization
# ================================
matplotlib>=3.5.0

# ================================
# Async Processing & Progress Tracking
# ================================
tqdm>=4.64.0

# ================================
# JSON Parsing & Repair
# ================================
json-repair>=0.7.0

# ================================
# HTTP Client (for API testing)
# ================================
requests>=2.28.0
